Created area_calc function and added import math to the equal distant point generator
The area_calc file is new, and being worked on. Currently the points have been sorted in the order that they appear going counterclockwise. Next the distances between the points can be calculated to find the area. Also added import math to the gen_equispace_pts file because it would not run without it.